add CLI arg for max-pending-msgs and send-queue-siz

+++ snow/networking/router/service_queue_test.go
@@ -18,7 +18,7 @@ import (
 )
 
 // returns a new multi-level queue that will never throttle or prioritize
-func setupMultiLevelQueue(t *testing.T, bufferSize int) (messageQueue, chan struct{}) {
+func setupMultiLevelQueue(t *testing.T, bufferSize uint32) (messageQueue, chan struct{}) {
 	metrics := &metrics{}
 	if err := metrics.Initialize("", prometheus.NewRegistry()); err != nil {
 		t.Fatal(err)
@@ -54,10 +54,10 @@ func setupMultiLevelQueue(t *testing.T, bufferSize int) (messageQueue, chan stru
 }
 
 func TestMultiLevelQueueSendsMessages(t *testing.T) {
-	bufferSize := 8
+	bufferSize := uint32(8)
 	queue, semaChan := setupMultiLevelQueue(t, bufferSize)
 	messages := []message{}
-	for i := 0; i < bufferSize; i++ {
+	for i := uint32(0); i < bufferSize; i++ {
 		messages = append(messages, message{
 			validatorID: ids.NewShortID([20]byte{byte(i)}),
 		})
@@ -67,7 +67,7 @@ func TestMultiLevelQueueSendsMessages(t *testing.T) {
 		queue.PushMessage(msg)
 	}
 
-	for count := 0; count < bufferSize; count++ {
+	for count := uint32(0); count < bufferSize; count++ {
 		select {
 		case _, ok := <-semaChan:
 			if !ok {
@@ -89,12 +89,12 @@ func TestMultiLevelQueueSendsMessages(t *testing.T) {
 }
 
 func TestExtraMessageNoDeadlock(t *testing.T) {
-	bufferSize := 8
+	bufferSize := uint32(8)
 	oversizedBuffer := bufferSize * 2
 	queue, semaChan := setupMultiLevelQueue(t, bufferSize)
 
 	messages := []message{}
-	for i := 0; i < oversizedBuffer; i++ {
+	for i := uint32(0); i < oversizedBuffer; i++ {
 		messages = append(messages, message{
 			validatorID: ids.NewShortID([20]byte{byte(i)}),
 		})
@@ -111,7 +111,7 @@ func TestExtraMessageNoDeadlock(t *testing.T) {
 	// because there is less than [bufferSize] room on the multi-level
 	// queue as a result of rounding when calculating the size of the
 	// single-level queues.
-	for i := 0; i < bufferSize; i++ {
+	for i := uint32(0); i < bufferSize; i++ {
 		<-semaChan
 	}
 	select {
@@ -122,7 +122,7 @@ func TestExtraMessageNoDeadlock(t *testing.T) {
 }
 
 func TestMultiLevelQueuePrioritizes(t *testing.T) {
-	bufferSize := 8
+	bufferSize := uint32(8)
 	vdrs := validators.NewSet()
 	validator1 := validators.GenerateRandomValidator(2000)
 	validator2 := validators.GenerateRandomValidator(2000)
@@ -235,7 +235,7 @@ func TestMultiLevelQueuePrioritizes(t *testing.T) {
 }
 
 func TestMultiLevelQueuePushesDownOldMessages(t *testing.T) {
-	bufferSize := 16
+	bufferSize := uint32(16)
 	vdrs := validators.NewSet()
 	vdr0 := validators.GenerateRandomValidator(2000)
 	vdr1 := validators.GenerateRandomValidator(2000)
@@ -336,7 +336,7 @@ func TestMultiLevelQueuePushesDownOldMessages(t *testing.T) {
 }
 
 func TestMultiLevelQueueFreesSpace(t *testing.T) {
-	bufferSize := 8
+	bufferSize := uint32(8)
 	vdrs := validators.NewSet()
 	validator1 := validators.GenerateRandomValidator(2000)
 	validator2 := validators.GenerateRandomValidator(2000)
@@ -394,7 +394,7 @@ func TestMultiLevelQueueFreesSpace(t *testing.T) {
 		metrics,
 	)
 
-	for i := 0; i < 4; i++ {
+	for i := uint32(0); i < 4; i++ {
 		validator1.ID()
 		if success := queue.PushMessage(message{
 			validatorID: validator1.ID(),
@@ -409,7 +409,7 @@ func TestMultiLevelQueueFreesSpace(t *testing.T) {
 	}
 
 	// Empty the message pool
-	for i := 0; i < bufferSize; i++ {
+	for i := uint32(0); i < bufferSize; i++ {
 		<-semaChan
 		if _, err := queue.PopMessage(); err != nil {
 			t.Fatalf("Failed to pop message on iteration %d due to: %s", i, err)
@@ -419,7 +419,7 @@ func TestMultiLevelQueueFreesSpace(t *testing.T) {
 
 	// Fill up message pool again to ensure
 	// popping previous messages freed up space
-	for i := 0; i < 4; i++ {
+	for i := uint32(0); i < 4; i++ {
 		if success := queue.PushMessage(message{
 			validatorID: validator1.ID(),
 		}); !success {
@@ -434,7 +434,7 @@ func TestMultiLevelQueueFreesSpace(t *testing.T) {
 }
 
 func TestMultiLevelQueueThrottles(t *testing.T) {
-	bufferSize := 8
+	bufferSize := uint32(8)
 	vdrs := validators.NewSet()
 	validator1 := validators.GenerateRandomValidator(2000)
 	validator2 := validators.GenerateRandomValidator(2000)
